Refactor project import dialog
