+++ asreview/webapp/src/Components/index.js
@@ -4,7 +4,7 @@ export { default as ExportDialog } from "./ExportDialog";
 export { default as Header } from "./Header";
 export { default as HistoryDialog } from "./HistoryDialog";
 export { default as HistoryListCard } from "./HistoryListCard";
-export { default as ImportDialog } from "./ImportDialog";
 export { default as MenuDrawer } from "./MenuDrawer";
 export { default as OpenInNewIconStyled } from "./OpenInNewIconStyled";
+export { default as ProjectImportDialog } from "./ProjectImportDialog";
 export { default as QuickTourDialog } from "./QuickTourDialog";
